requirements.txt
bcrypt==5.0.0
fastapi==0.128.8
fastapi-cli==0.0.22
fastapi-cloud-cli==0.12.0
jwt==1.4.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.2
psycopg2-binary==2.9.11
SQLAlchemy==2.0.46
uvicorn==0.39.0
qrcode
pillow
Authlib==1.6.8
python-jose==3.5.0
itsdangerous==2.2.0
